requirements.txt
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
pyautogui>=0.9.54
keyboard>=0.13.5
pywin32>=306
pystray>=0.19.5
pillow>=10.0.0
pyinstaller>=6.0.0
